Use TextActionFragment for TextAction and TextActionSet
