+++ app/src/androidTest/java/com/hedvig/app/feature/embark/screens/TextActionSetScreen.kt
@@ -6,13 +6,13 @@ import com.agoda.kakao.recycler.KRecyclerItem
 import com.agoda.kakao.recycler.KRecyclerView
 import com.agoda.kakao.text.KButton
 import com.hedvig.app.R
-import com.hedvig.app.feature.embark.passages.textactionset.TextActionSetFragment
+import com.hedvig.app.feature.embark.passages.textactionset.TextActionFragment
 import com.kaspersky.kaspresso.screens.KScreen
 import org.hamcrest.Matcher
 
 object TextActionSetScreen : KScreen<TextActionSetScreen>() {
     override val layoutId = R.layout.fragment_text_action_set
-    override val viewClass = TextActionSetFragment::class.java
+    override val viewClass = TextActionFragment::class.java
 
     val submit = KButton { withId(R.id.textActionSubmit) }
