+++ app/src/main/java/com/hedvig/app/feature/embark/ui/EmbarkActivity.kt
@@ -26,10 +26,8 @@ import com.hedvig.app.feature.embark.passages.previousinsurer.PreviousInsurerFra
 import com.hedvig.app.feature.embark.passages.previousinsurer.PreviousInsurerParameter
 import com.hedvig.app.feature.embark.passages.selectaction.SelectActionFragment
 import com.hedvig.app.feature.embark.passages.selectaction.SelectActionParameter
-import com.hedvig.app.feature.embark.passages.textaction.TextActionFragment
-import com.hedvig.app.feature.embark.passages.textaction.TextActionParameter
-import com.hedvig.app.feature.embark.passages.textactionset.TextActionSetFragment
-import com.hedvig.app.feature.embark.passages.textactionset.TextActionSetParameter
+import com.hedvig.app.feature.embark.passages.textactionset.TextActionFragment
+import com.hedvig.app.feature.embark.passages.textactionset.TextActionParameter
 import com.hedvig.app.feature.settings.MarketManager
 import com.hedvig.app.feature.settings.SettingsActivity
 import com.hedvig.app.feature.webonboarding.WebOnboardingActivity
@@ -181,12 +179,12 @@ class EmbarkActivity : BaseActivity(R.layout.activity_embark) {
 
         passage?.action?.asEmbarkTextActionSet?.let { textActionSet ->
             textActionSet.textSetData?.let { data ->
-                val parameter = TextActionSetParameter.from(
+                val parameter = TextActionParameter.from(
                     passage.messages.map { it.fragments.messageFragment.text },
                     data,
                     passage.name
                 )
-                return TextActionSetFragment.newInstance(parameter)
+                return TextActionFragment.newInstance(parameter)
             }
         }
